cmd/AppArmor moved to subpackage
