+++ cmd/nerdctl/apparmor/apparmor_unload_linux.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package apparmor
 
 import (
 	"fmt"
