+++ cmd/nerdctl/main_linux.go
@@ -19,6 +19,7 @@ package main
 import (
 	"github.com/spf13/cobra"
 
+	"github.com/containerd/nerdctl/v2/cmd/nerdctl/apparmor"
 	"github.com/containerd/nerdctl/v2/pkg/rootlessutil"
 	"github.com/containerd/nerdctl/v2/pkg/strutil"
 )
@@ -63,7 +64,7 @@ func appNeedsRootlessParentMain(cmd *cobra.Command, args []string) bool {
 }
 
 func addApparmorCommand(rootCmd *cobra.Command) {
-	rootCmd.AddCommand(newApparmorCommand())
+	rootCmd.AddCommand(apparmor.NewApparmorCommand())
 }
 
 func addCpCommand(rootCmd *cobra.Command) {
